% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey-plot.R
\name{sankey_plot}
\alias{sankey_plot}
\title{Create Sankey Plot}
\usage{
sankey_plot(sankey_data, colors, version = 1)
}
\arguments{
\item{sankey_data}{A data frame containing the melted data for the Sankey plot.}

\item{colors}{A named vector of colors for the different risk factors.}

\item{version}{An integer (1 or 2) specifying the version of the plot. Version 1 has vertical lines and horizontal white space, while version 2 has no white space.}
}
\value{
A ggplot object representing the Sankey plot.
}
\description{
This function creates a Sankey plot for visualizing the contribution of various risk factors over time.
}
\examples{
colors <- c("Hypertension" = "#94B992",
            "Diabetes" = "#CA6C56",
            "Smoking" = "#87AAC2",
            "Hypercholesterolemia" = "#CD925F",
            "Obesity" = "#546C91")
load(system.file("data", "sankey_data.rda", package = "ymsankey"))
sankey_plot(sankey_data, colors, version = 1)
}
